Update README to point out Kali's issues
After attempt to debug AFL and Panda for 2+ days the solution was finally found to the compilation issues...

For reference - 

https://patchwork.kernel.org/patch/8982051/
http://oss.sgi.com/archives/xfs/2016-04/msg00598.html

From this and checking my libraries it's clear Kali is running with some very outdated libraries.
To be fair it was never intended to be used day-to-day, but for pentesting purposes.

Issues with compilation with Kali is from experience because of this - so there's not much one can do.
Mucking around with libraries especially in Kali breaks some tools due to being outdated (then again some pentesting tools were written as PoC in the first place...). Dynamic Linking strikes again!